[Refactor] Serialise modules using a binary format
It turns out that simply printing them as source code and relying
on the parser for everything doesn’t really work for the AST as
soon as we have to deal with things that aren’t really represented
in source, e.g. the layout of a type imported from C—there is no way
to encode the offsets in source, and it’d honestly be rather stupid
to add an attribute or some other builtin just for modules.

Reusing the parser was supposed to avoid introducing special handling
for modules, but since it has turned out to require that anyway, just
revert to using a proper serialiser for this.